front/requirements.txt
aiogram>=3.0.0
environs==14.1.1
motor==3.7.0
python-dotenv>=1.0.0
httpx>=0.24.0